#130 Added support for google fit steps data streams
