+++ google_fit/get_refresh_token.py
@@ -1,16 +1,21 @@
+import sys
+import os
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
+from client.GoogleFitClient import GoogleFitClient
+
 import time
 import json
 
 from datetime import datetime
 
-import config
-from GoogleFitHeartRate import GoogleFitHeartRate
+import client.config as config
 
 # Open the refresh token file
 token_file = open("refresh_token", "w")
 
 # Get the refresh token
-refresh_token = GoogleFitHeartRate.get_refresh_token(config.CLIENT_ID, config.CLIENT_SECRET)
+refresh_token = GoogleFitClient.get_refresh_token(config.CLIENT_ID, config.CLIENT_SECRET)
 
 # Write the token to the file
 token_file.write(refresh_token)
